+++ src/Animals/Eagle.java
@@ -18,7 +18,7 @@ public class Eagle extends Bird implements IFly {
         return nbEagle;
     }
 
-    public Eagle(String name, boolean sex, int weight, int height, int age) {
+    public Eagle(String name, boolean sex, long weight, int height, int age) {
         super(name, sex, weight, height, age);
         timeUnborn = Tools.random(6000, 6301);// 1000 ms égale 1 jour en vrai
         typeAnimal = "volant";
