+++ src/Species/Mammal.java
@@ -6,7 +6,7 @@ package Species;
 public abstract class Mammal extends Animal {
 
 
-    protected Mammal(String name, boolean sex, int weight, int height, int age) {
+    protected Mammal(String name, boolean sex, long weight, int height, int age) {
         super(name, sex, weight, height, age);
     }
